feat: Add FitGirl feed section to the home tools feed
This commit adds a new feed section called FitGirl to the home tools feed. It includes the necessary changes to the index.tsx file to import the FitGirl component and display it when the source is set to "fitgirl". The FitGirl component fetches data from a JSON file and renders a list of posts with relevant information.